requirements.txt
numpy==1.24.3
opencv_python==4.5.5.64
Pillow==9.2.0
Pillow==9.5.0
streamlit==1.23.1
